Trying to fix minor bug in follows
